Sync V2: build Scan QR screen with Scan/Enter Code tabs
Add ScanQRCodeViewV2 with a segmented Scan QR / Enter Code control, the
Enter Code tab (EnterCodeTabView), and the scan tab (ScanTabView). Extract
reusable SyncAppNameChip and SyncInstructionText, wire them into the code
sheet and enter-code views, and add rebranded previews.
