+++ iOS/LocalPackages/SyncUI-iOS/Sources/SyncUI-iOS/Views/Internal/UserText.swift
@@ -176,9 +176,13 @@ public struct UserText {
     static let simplifiedScanTabEnterCode = NotLocalizedString("sync.simplified.scan.v2.tab.enter.code", bundle: Bundle.module, value: "Enter Code", comment: "V2 segmented control tab title for entering a code")
     static let simplifiedSyncCodeSheetTitle = NotLocalizedString("sync.simplified.sync.code.sheet.title", bundle: Bundle.module, value: "Scan or Copy Code", comment: "Navigation title of the sheet that shows this device's sync QR code")
     static let simplifiedSyncCodeSheetOpenInstruction = NotLocalizedString("sync.simplified.sync.code.sheet.open", bundle: Bundle.module, value: "On another device, open", comment: "First part of the sync-code sheet instructions, followed by a DuckDuckGo app chip")
-    static let simplifiedSyncCodeAppName = NotLocalizedString("sync.simplified.sync.code.app.name", bundle: Bundle.module, value: "DuckDuckGo", comment: "DuckDuckGo app name shown in the app chip on the sync-code sheet")
     static let simplifiedSyncCodeSheetScanInstruction = NotLocalizedString("sync.simplified.sync.code.sheet.scan", bundle: Bundle.module, value: "Go to **Sync & Backup** › **Sync With Another Device** and scan this code:", comment: "Second part of the sync-code sheet instructions. **...** marks the parts shown in the primary text color.")
     static let simplifiedSyncCodeCopyButton = NotLocalizedString("sync.simplified.sync.code.copy.button", bundle: Bundle.module, value: "Copy Text Code", comment: "Button to copy the text sync code on the sync-code sheet")
+    static let simplifiedEnterCodeTitle = NotLocalizedString("sync.simplified.enter.code.title", bundle: Bundle.module, value: "Enter the Sync Code", comment: "Title on the enter-code tab of the V2 scan screen")
+    static let simplifiedEnterCodeOpenInstruction = NotLocalizedString("sync.simplified.enter.code.open", bundle: Bundle.module, value: "Open your other", comment: "First part of the enter-code instructions, followed by a DuckDuckGo app chip")
+    static let simplifiedEnterCodeStepsInstruction = NotLocalizedString("sync.simplified.enter.code.steps", bundle: Bundle.module, value: "Go to **Sync & Backup** › **Sync With Another Device** › **Copy Text Code**.", comment: "Second part of the enter-code instructions. **...** marks the parts shown in the primary text color.")
+    static let simplifiedEnterCodeExampleLabel = NotLocalizedString("sync.simplified.enter.code.example.label", bundle: Bundle.module, value: "Example Code:", comment: "Label above the example sync code placeholder on the enter-code tab")
+    static let simplifiedEnterCodePasteButton = NotLocalizedString("sync.simplified.enter.code.paste.button", bundle: Bundle.module, value: "Paste Sync Code", comment: "Button to paste the sync code on the enter-code tab")
     static let simplifiedScanInstructions = NSLocalizedString("sync.simplified.scan-or-view-code.scan.instructions.line1", bundle: Bundle.module, value: "Open the DuckDuckGo app on your other device.", comment: "First line of QR code scanning instructions")
     static let simplifiedScanInstructionsLine2 = NSLocalizedString("sync.simplified.scan-or-view-code.scan.instructions.line2", bundle: Bundle.module, value: "Find the QR code in Settings > Sync & Backup > Sync With Another Device.", comment: "Second line of QR code scanning instructions")
     static let simplifiedScanCameraPrompt = NSLocalizedString("sync.simplified.scan-or-view-code.camera.prompt", bundle: Bundle.module, value: "Point Camera at QR to Scan", comment: "Floating prompt over QR code scanner preview")
